Extract managed thread runtime service
