+++ src/codex_autorunner/surfaces/cli/commands/hub.py
@@ -39,7 +39,7 @@ from ....core.orchestration.sqlite import open_orchestration_sqlite
 from ....core.pma_automation_store import PmaAutomationStore
 from ....manifest import Manifest, load_manifest, save_manifest
 from ...web.app import create_hub_app
-from ...web.routes.pma_routes.managed_thread_runtime import (
+from ...web.services.pma.managed_thread_runtime import (
     recover_orphaned_managed_thread_executions,
 )
 from ..hub_path_option import hub_root_path_option
