+++ src/codex_autorunner/surfaces/web/routes/pma_routes/__init__.py
@@ -1,16 +1,20 @@
 from __future__ import annotations
 
 from ...services.pma.automation import notify_managed_thread_terminal_transition
+from . import managed_thread_runtime as _managed_thread_runtime
 from .action_manifest import build_action_manifest_routes
 from .chat_events import build_chat_event_routes
 from .chat_runtime import build_chat_runtime_router
 from .history_files_docs import build_history_files_docs_router
-from .managed_thread_runtime import build_managed_thread_runtime_routes
 from .managed_threads import build_automation_routes, build_managed_thread_crud_routes
 from .meta import build_pma_meta_routes
 from .publish import publish_automation_result
 from .runtime_state import PmaRuntimeState
 from .tail_stream import build_managed_thread_tail_routes
+
+build_managed_thread_runtime_routes = (
+    _managed_thread_runtime.build_managed_thread_runtime_routes
+)
 
 __all__ = [
     "PmaRuntimeState",
